Add GDAL option and create portable bundle
* Allow compiling without GDAL
* Create portable bundle on Windows
